backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
transformers==4.40.0
torch>=2.1.0
sentencepiece==0.1.99
peft==0.10.0
accelerate==0.29.3
cachetools==5.3.2
python-dotenv==1.0.0
